refactor: remove interfaces
