--- src/@types/express.d.ts
@@ -1,8 +1,0 @@
-declare namespace Express {
-  export interface Request {
-    user: {
-      id: number;
-      access_level: number;
-    };
-  }
-}
