+++ src/shared/providers/DiskStorage.ts
@@ -1,11 +1,9 @@
 import fs from 'fs';
 import path from 'path';
 
-import uploadConfig from '../../../../config/Multer';
+import uploadConfig from '../../config/Multer';
 
-import IStorageProvider from '../model/IStorageProvider';
-
-export default class DiskStorageProvider implements IStorageProvider {
+export default class DiskStorageProvider {
   public async saveFile(file: string): Promise<string> {
     await fs.promises.rename(
       path.resolve(uploadConfig.tmpFolder, file),
