--- src/shared/providers/MailProvider/model/IMailProvider.ts
@@ -1,5 +1,0 @@
-import ISendMailDTO from '../dtos/ISendMailDTO';
-
-export default interface IMailProvider {
-  sendMail(data: ISendMailDTO): Promise<void>;
-}
