--- src/shared/providers/MailTemplateProvider/models/IMailTemplateProvider.ts
@@ -1,5 +1,0 @@
-import IParserMailTemplateDTO from '../dtos/IParserMailTemplateDTO';
-
-export default interface IMailTemplateProvider {
-  parse(data: IParserMailTemplateDTO): Promise<string>;
-}
