+++ src/shared/providers/dtos/IParserMailTemplate.ts
