--- src/shared/providers/storageProvider/model/IStorageProvider.ts
@@ -1,4 +1,0 @@
-export default interface IStorageProvider {
-  saveFile(file: string): Promise<string>;
-  deleteFile(file: string): Promise<void>;
-}
